Fix weird loading error of prog textures, causing distortion

+++ source/glquake.h
@@ -51,7 +51,7 @@ extern	int		texture_mode;
 
 extern	float	gldepthmin, gldepthmax;
 
-void GL_Upload32 (byte *data, int width, int height,  qboolean mipmap, qboolean alpha);
+void GL_Upload32 (unsigned *data, int width, int height,  qboolean mipmap, qboolean alpha);
 void GL_Upload8 (byte *data, int width, int height,  qboolean mipmap, qboolean alpha);
 int GL_LoadTexture (char *identifier, int width, int height, byte *data, qboolean mipmap, qboolean alpha, int bytesperpixel);
 int GL_FindTexture (char *identifier);
